requirements.txt
alembic==1.4.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-editor==1.0.4
Flask-RESTful==0.3.8
Flask-Cors==3.0.8
rsa==4.0
gunicorn==20.0.4
SQLAlchemy==1.3.17
python-jose~=3.1.0